Add scripts to add deps and replace SlackBuilds
Add sb_adddeps script that adds new dependencies to the REQUIRES list of
SlackBuilds. For example, add "qt" as a dependency to skype.

Add sb_replace script that replaces SlackBuild names with new names, in
case the package is already present in Salix repos, but is named
differently. For example "configobj" and "python-configobj".

In sb_dupes_req, also exclude all SlackBuilds in the SBoREPLACE list.